chore: move resolvers in separate folders
